Create functions to check for victory conditions
Have created functions to check whether any given square has a matching value in an adjacent square. These functions behave appropriately when directly called in console with a specific square (e.g. checkRight(1,1) will return true if there are X's in both 1,1 and 1,2), but the functions do not play nice in the overall detectWin() function yet.