Create working detectWin function
The function detectWin takes currentBoard as a parameter. It calls three functions sequentially, the first to check if a player has won a row, the second to check if a player has won a column, and the third to check if a player has won a diagonal. Each of those functions checks to see if each of their respective possible combinations has won, then verifies that moves has actually been made on those squares (the detectWinCol function would originally return true for a blank board since all three of the values it compared would always been equal at the start of the game). This function does not yet determine which player won, and has not been integrated into the rest of the game logic.