Fix victory check logic
The game was not correctly determining victory whenever a player won by claiming either the right-hand column or the down-to-the-left diagonal. The first problem was caused by the _detectWinCol function terminating its loop before checking the right column. The second problem was caused by the _detectWinDiag function incorrectly returning the contents of the top-left square when it should have returned the contents of the top-right square. These have both been corrected.